+++ NEWS.md
@@ -9,6 +9,9 @@
 
 * Removed `work_only` argument from `file_outline()`, use `pattern = "WORK"` to 
   acheive the same result.
+  
+* In `proj_outline()`, `proj` has been renamed `path`, but still accepts a project name,
+that will passed on to `proj_list()`
 
 ## Fixes